Moved location of exported header files so that external projects may include them more easily
