+++ CMakeLists.txt
@@ -27,7 +27,3 @@ set (UMAPINCLUDEDIRS
 add_subdirectory(src)
 add_subdirectory(examples)
 
-if ( ENABLE_TESTS )
-  add_subdirectory(tests)
-endif()
-
