+++ src/store/Store.cpp
@@ -13,8 +13,8 @@
  * the Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
  * 02111-1307 USA
  */
-#include "umap.h"
-#include "Store.h"
+#include "umap/umap.h"
+#include "umap/Store.h"
 #include "StoreFile.h"
 
 using namespace std;
