+++ tests/umapcpu/umapcpu.cpp
@@ -42,7 +42,7 @@
 #include <omp.h>
 #endif
 
-#include "umap.h"
+#include "umap/umap.h"
 #include "../utility/commandline.hpp"
 #include "../utility/umap_file.hpp"
